project/backend/requirements.txt
# Flask Backend Dependencies - 2024 Optimized for Google Cloud Run

# Core Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1

# CORS and Security
Flask-CORS==4.0.0
Flask-WTF==1.2.1

# Google Cloud Dependencies
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-cloud-sql-connector==1.4.3

# RSS and Web Processing
feedparser==6.0.10
requests==2.31.0
html2text==2020.1.16
beautifulsoup4==4.12.2
lxml==4.9.3

# EPUB and PDF Processing
EbookLib==0.18
PyPDF2==3.0.1
Pillow==10.1.0

# Email Support
secure-smtplib==0.1.1

# Data Processing
python-dateutil==2.8.2
pytz==2023.3
regex==2023.10.3

# Caching and Performance
redis==5.0.1
Flask-Caching==2.1.0

# Rate Limiting
Flask-Limiter==3.5.0

# Task Queue (for background processing)
celery==5.3.4
celery[redis]==5.3.4

# Monitoring and Logging
structlog==23.2.0
python-json-logger==2.0.7

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Production WSGI Server
gunicorn==21.2.0

# Environment and Configuration
python-dotenv==1.0.0
click==8.1.7

# Utilities
shortuuid==1.0.11
python-magic==0.4.27
urllib3==2.1.0

# Type Hints
typing-extensions==4.8.0

# JSON Web Tokens (if authentication needed)
PyJWT==2.8.0

# HTTP Client with retries
httpx==0.25.2

# XML Processing
xmltodict==0.13.0

# Timezone handling
babel==2.13.1

# File format detection
python-magic-bin==0.4.14

# Additional text processing
nltk==3.8.1
textstat==0.7.3